Upgrade to Laravel 8
This moves us to L8. Most of the work has been put into making the tests pass, as we have moved to using the new, "improved" factories, rather than add the legacy factories package.

Additionally we have moved models to a `Models` namespace, to be in line with how a new L8 project is structured in that regard.