Update for use of real facade
The Facades\Sendportal class was not correctly set up as a facade. It is unclear about how it was working before, but it stopped. The core package now correctly implements the facade, and additionally renames some of the methods to be clearer about what they do.

This commit ensures that we use the facade from the core package correctly, largely by updating to use the new method names.

Additionally, some minor adjustments have been made to the implementations of things like the workspace ID resolver, to make them a tiny bit more robust (checking for null and that sort of thing).