Add workspace ID resolution from `api_token` parameter
The previous version of SP included the ability to use API tokens as either a Bearer token, or supplied as an `api_token` parameter in a request (GET or POST). This new packaged version removed that ability, but the various middlewares still let it through as a valid token, ending up hitting us with a RuntimeException thrown from the workspace ID resolver.

We now check for the API token in the resolver bother as a bearer token and as an `api_token` parameter, and subsequently avoid seeing the RuntimeException in this case, with the advantage of keeping a little backward compatibility with v1.