Automatically accept invitations to existing users
This was prompted by an issue where users who are only in the system
via an invitation, who subsequently get removed from the workspace they
were invited to, were unable to access the application at all to either
create their own workspace or accept an invitation to another existing
one.

By automatically accepting invitations that are made to existing users,
we avoid having to do things like deleting users if they have no
workspace, or adjusting the system so that it workspaceless users can
access workspace management features to create or accept invitations to
workspaces.

Additionally, this removes friction from adding users to new workspaces
in any given SendPortal instance. It seems unlikely that someone would
actually want to reject an invitation, given that it is a self-hosted
platform.

The ability to manually accept invitations has not been removed, as
there may be cases where invitations already exist that haven't yet
been handled.